+++ hutool-core/src/main/java/cn/hutool/core/io/FastByteArrayOutputStream.java
@@ -52,7 +52,7 @@ public class FastByteArrayOutputStream extends OutputStream {
 	 * 此方法无任何效果，当流被关闭后不会抛出IOException
 	 */
 	@Override
-	public void close() {
+	public void close() throws IOException{
 		// nop
 	}