+++ hutool-extra/src/main/java/cn/hutool/extra/ftp/FtpMode.java
@@ -1,0 +1,17 @@
+package cn.hutool.extra.ftp;
+
+/**
+ * FTP连接模式
+ * 
+ * <p>
+ * 见：https://www.cnblogs.com/huhaoshida/p/5412615.html
+ * 
+ * @author looly
+ * @since 4.1.19
+ */
+public enum FtpMode {
+	/** 主动模式 */
+	Active,
+	/** 被动模式 */
+	Passive;
+}